Changed AutoCriticalSection to Loki::ScopeGuard.
git-svn-id: http://arkanar/svn/Perimeter%20Map%20Compiler@71 e62d0a23-8ff8-fb48-9cb0-22a7500eb68d

+++ trunk/compiler/src/task common.cpp
@@ -601,7 +601,7 @@ namespace TaskCommon
 			for (LONG r(0); r != map_size.cy; ++r)
 			{
 				for (LONG c(0); c != map_size.cx; ++c)
-					*img_iter++ = *ihm_iter++ << 3;
+					*img_iter++ = static_cast<WORD>(*ihm_iter++ << 3);
 				++ihm_iter;
 			}
 			img.save("pre_blur.tiff", TIFF_NONE);
